partition files and make address relocate done
