+++ lab_2/bootloader/utils.c
