--- lab_2/config.txt
@@ -1,3 +1,0 @@
-kernel_address=0x60000
-kernel=kernel8.img
-arm_64bit=1
